stream: move event filtering to PayloadEvents
Removes the weirdness around PayloadEvents.FilterByKey
